+++ intelmq/tests/lib/test_message.py
@@ -615,5 +615,18 @@ class TestMessageFactory(unittest.TestCase):
             message.Event(harmonization={'event': {'foo.bar.': {}}})
 
 
+class TestReport(unittest.TestCase):
+    """
+    Test the Report class.
+    """
+    def test_report_from_event(self):
+        event = message.Event(harmonization=HARM)
+        event.add('feed.code', 'adasd')
+        event.add('source.fqdn', 'example.com')
+        report = message.Report(event).to_dict()
+        self.assertNotIn('source.fqdn', report)
+        self.assertIn('feed.code', report)
+
+
 if __name__ == '__main__':  # pragma: no cover  # pragma: no cover
     unittest.main()